+++ src/components/FormComponents/PrimaryInfoForm/PrimaryInfoForm.tsx
@@ -6,7 +6,7 @@ import Select from "@nepMeds/components/Form/Select";
 import { useGetDistricts, useGetProvince } from "@nepMeds/service/nepmeds-core";
 import { useSpecializationData } from "@nepMeds/service/nepmeds-specialization";
 import { colors } from "@nepMeds/theme/colors";
-import { district, gender, idType, phone } from "@nepMeds/utils/choices";
+import { gender, idType, phone } from "@nepMeds/utils/choices";
 import { useFormContext } from "react-hook-form";
 import { IRegisterFields } from "../RegistrationForm/RegistrationForm";
 import { IGetDoctorProfile } from "@nepMeds/service/nepmeds-doctor-profile";
@@ -62,6 +62,7 @@ const PrimaryInfo = ({
           label="Mobile No."
           name="mobile_number"
           type="tel"
+          required
           register={register}
           defaultValue={doctorProfileData?.user?.mobile_number}
           style={{ background: colors.forminput, border: "none" }}
@@ -71,6 +72,7 @@ const PrimaryInfo = ({
         <FloatingLabelInput
           type="email"
           label="Email"
+          required
           name="email"
           register={register}
           defaultValue={doctorProfileData?.user?.email}
@@ -105,6 +107,7 @@ const PrimaryInfo = ({
       <GridItem colSpan={2}>
         <MultiSelect
           label="Specialization"
+          required
           name="specialization"
           multiValue={doctorProfileData?.specialization.map(item => ({
             label: item,
@@ -123,6 +126,7 @@ const PrimaryInfo = ({
         <FloatingLabelInput
           label="Pan Number"
           name="pan_number"
+          required
           register={register}
           style={{
             background: colors.forminput,
@@ -161,11 +165,12 @@ const PrimaryInfo = ({
         <Select
           placeholder=""
           label="Issued District"
-          name="issued_district"
+          name="citizenship_issued_district"
           required
+          defaultValue=""
           register={register}
+          options={districtOptions}
           defaultValue={doctorProfileData?.citizenship_issued_district}
-          options={district}
           style={{
             background: colors.forminput,
             border: "none",
@@ -246,6 +251,7 @@ const PrimaryInfo = ({
           placeholder=""
           label="Tole"
           name="tole"
+          required
           register={register}
           defaultValue={doctorProfileData?.tole}
           style={{ background: colors.forminput, border: "none" }}