+++ src/components/Table/Doctor/ApprovedDocList.tsx
@@ -60,6 +60,12 @@ const ApprovedDocList = () => {
           return (
             <Badge
               colorScheme={profile_status === "approved" ? "green" : "yellow"}
+              p={1}
+              borderRadius={20}
+              fontSize={11}
+              w={20}
+              textAlign="center"
+              textTransform="capitalize"
             >
               {profile_status}
             </Badge>
@@ -90,7 +96,7 @@ const ApprovedDocList = () => {
   const { data, isLoading } = useApprovedDoctorList();
   const [searchFilter, setSearchFilter] = useState("");
   const [id, setId] = React.useState("");
-  const { data: detail } = useDoctorDetail(id);
+  const { data: detail, isLoading: isFetching } = useDoctorDetail(id);
   if (isLoading)
     return (
       <Spinner
@@ -123,7 +129,7 @@ const ApprovedDocList = () => {
       />
       <ModalComponent
         alignment="left"
-        size="6xl"
+        size="3xl"
         isOpen={isDetailsModalOpen}
         onClose={onDetailsModalClose}
         heading={
@@ -134,7 +140,13 @@ const ApprovedDocList = () => {
         }
         footer={<></>}
       >
-        <DoctorDetail {...detail} />
+        {isFetching ? (
+          <Spinner
+            style={{ margin: "0 auto", textAlign: "center", display: "block" }}
+          />
+        ) : (
+          <DoctorDetail {...detail} />
+        )}
       </ModalComponent>
     </>
   );