+++ src/routes/AppRoutes.tsx
@@ -12,10 +12,15 @@ import CertificationInfo from "@nepMeds/pages/Register/CertificationInfo";
 import ExperienceInfo from "@nepMeds/pages/Register/ExperienceInfo";
 import PrimaryInfo from "@nepMeds/pages/Register/PrimaryInfo";
 import SignUp from "@nepMeds/pages/SignUp/SignUp";
-import { useAuthentication } from "@nepMeds/service/nepmeds-auth";
+import {
+  useAuthentication,
+  useLoginTokenDetailQuery,
+} from "@nepMeds/service/nepmeds-auth";
 import { Navigate, useRoutes } from "react-router-dom";
 import { NAVIGATION_ROUTES } from "./routes.constant";
 import DoctorProfile from "@nepMeds/pages/DoctorList/DoctorProfile";
+import { Suspense } from "react";
+import { Center, Spinner } from "@chakra-ui/react";
 
 const routes = [
   {
@@ -50,6 +55,10 @@ const routes = [
         path: NAVIGATION_ROUTES.PAYMENT,
         element: <>Payment</>,
       },
+      {
+        path: NAVIGATION_ROUTES.DOCTOR_PROFILE,
+        element: <DoctorProfile />,
+      },
     ],
   },
 ];
@@ -70,10 +79,7 @@ const adminRoutes = [
         path: NAVIGATION_ROUTES.DOCTOR_LIST,
         element: <AllDoctors />,
       },
-      {
-        path: NAVIGATION_ROUTES.DOCTOR_PROFILE,
-        element: <DoctorProfile />,
-      },
+
       {
         path: NAVIGATION_ROUTES.PATIENTS,
         element: <>Patients</>,
@@ -157,14 +163,28 @@ const openRoutes = [
 ];
 
 const AppRoutes = () => {
-  const { data: isAuthenticated } = useAuthentication();
-  // const dataInfo = useUserInfoQuery();
-  // const is_doctor = dataInfo.data?.data?.data?.is_doctor;
-  // localStorage.setItem("doctor", is_doctor);
-  return useRoutes(
-    isAuthenticated ? adminRoutes : openRoutes
+  const { data: isAuthenticated, isLoading } = useAuthentication();
+  const { data: userInfo } = useLoginTokenDetailQuery();
 
-    // isAuthenticated ? (is_doctor ? routes : adminRoutes) : openRoutes
+  if (isLoading) {
+    return (
+      <Center h="100vh">
+        <Spinner />
+      </Center>
+    );
+  }
+
+  return (
+    <Suspense fallback={<Spinner />}>
+      {useRoutes(
+        // isAuthenticated ? adminRoutes : openRoutes
+        isAuthenticated
+          ? userInfo?.is_superuser
+            ? adminRoutes
+            : routes
+          : openRoutes
+      )}
+    </Suspense>
   );
 };