+++ src/service/nepmeds-auth.ts
@@ -17,17 +17,8 @@ export interface LoginDetails {
   email: string;
   password: string;
 }
-type AuthInfo = {
-  first_name: string;
-  last_name: string;
-  email: string;
-  mobile_number: string;
-  is_doctor: boolean;
-  is_superuser: boolean;
-};
-
 export const authTokenKey = "authToken";
-export const auth = "authInfo";
+export const auth = "userInfo";
 const authTokenDetails = "authTokenDetails";
 
 const initLogout = () => {
@@ -47,6 +38,8 @@ const useLogoutMutation = (noToast?: boolean) => {
       logoutChannel.postMessage("Logout");
       queryClient.clear();
       queryClient.setQueryData(authTokenKey, () => false);
+      localStorage.setItem("doctor", "false");
+      localStorage.setItem("admin", "false");
       navigate("/", { replace: true });
       !noToast && toastSuccess("Logged out Succesfully");
     },
@@ -60,7 +53,6 @@ const initLogin = (loginData: LoginDetails) => {
 const useLoginMutation = () => {
   const queryClient = useQueryClient();
   const navigate = useNavigate();
-
   return useMutation(initLogin, {
     onSuccess: response => {
       loginChannel.postMessage(loginBroadcast);
@@ -70,9 +62,8 @@ const useLoginMutation = () => {
       };
       TokenService.setToken(tokens);
       queryClient.setQueryData(authTokenKey, () => true);
-      queryClient.setQueryData(auth, () => response);
+      navigate("/dashboard");
       toastSuccess("Login Successful!!");
-      navigate("/dashboard", { replace: true });
     },
     onError: error => {
       const loginErr = error as AxiosError<{ message: string; error: string }>;
@@ -85,38 +76,42 @@ const useLoginMutation = () => {
   });
 };
 
-const initRefreshToken = async () => {
-  try {
-    // const response = await HttpClient.get<TokenDetails>(api.refreshToken, {
-    //   params: {
-    //     refreshToken: TokenService.getToken()?.refresh_token,
-    //   },
-    //   headers: {
-    //     Authorization: "",
-    //   },
-    // });
-    // const tokens = {
-    //   access_token: response.data.access_token,
-    //   refresh_token: response.data.refresh_token,
-    // };
-    // TokenService.setToken(tokens);
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
+// const initRefreshToken = async () => {
+//   try {
+//     // const response = await HttpClient.get<TokenDetails>(api.refreshToken, {
+//     //   params: {
+//     //     refreshToken: TokenService.getToken()?.refresh_token,
+//     //   },
+//     //   headers: {
+//     //     Authorization: "",
+//     //   },
+//     // });
+//     // const tokens = {
+//     //   access_token: response.data.access_token,
+//     //   refresh_token: response.data.refresh_token,
+//     // };
+//     // TokenService.setToken(tokens);
+//     return true;
+//   } catch (error) {
+//     return false;
+//   }
+// };
 
 const checkAuthentication = async () => {
-  if (TokenService.isAuthenticated()) {
-    const tokenInfo = TokenService.getTokenDetails();
-    if (tokenInfo && tokenInfo.exp * 1000 < Date.now() + 5 * 60 * 1000) {
-      return initRefreshToken();
-    }
+  // if (TokenService.isAuthenticated()) {
+  //   const tokenInfo = TokenService.getTokenDetails();
+  //   if (tokenInfo && tokenInfo.exp * 1000 < Date.now() + 5 * 60 * 1000) {
+  //     return initRefreshToken();
+  //   }
+  //   return Promise.resolve(true);
+  // } else if (TokenService.getToken()?.refresh) {
+  //   return initRefreshToken();
+  // }
+  // return Promise.resolve(null);
+  if (TokenService.getTokenDetails()) {
     return Promise.resolve(true);
-  } else if (TokenService.getToken()?.refresh) {
-    return initRefreshToken();
   }
-  return Promise.resolve(null);
+  return Promise.resolve(false);
 };
 
 /**
@@ -138,8 +133,12 @@ const useAuthentication = () => {
   });
 };
 
+const checkUserInfo = async () => {
+  return TokenService.getTokenDetails();
+};
+
 export const useUserInfoQuery = () => {
-  return useQuery<AuthInfo>(auth);
+  return useQuery<TokenInfo | null>(auth, checkUserInfo);
 };
 
 const useLoginTokenDetailQuery = () => {